fixed bug in dataloader; updated copiot training loss to be on a log scale; updated lunar lander human-in-the-loop game to have a timer; also updated human-in-the-loop experiments so it is impossible to stamp over existing directories; performed refactor of benchmarking script to allow compatability with both lunar lander and reacher environments

+++ copilots/dataset.py
@@ -21,7 +21,7 @@ class ExpertDemonstrations(Dataset):
         if torch.is_tensor(idx):
             idx = idx.tolist()
         expert_demonstrations = self.expert_demonstrations[idx]
-        return expert_demonstrations
+        return np.float32(expert_demonstrations)
     
 if __name__ == '__main__':
     dataset = ExpertDemonstrations('/home/necl/code/lunar-lander/expert_trajectories_random_goal.npy')
